+++ axum/CHANGELOG.md
@@ -56,6 +56,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
         - `MatchedPathRejection`
         - `WebSocketUpgradeRejection`
 - **breaking:** `Redirect::found` has been removed ([#800])
+- **breaking:** `AddExtensionLayer` has been removed. Use `Extension` instead. It now implements
+  `tower::Layer`
 - **fixed:** Set `Allow` header when responding with `405 Method Not Allowed` ([#733])
 - **fixed:** Correctly set the `Content-Length` header for response to `HEAD`
   requests ([#734])